+++ entwuerfe/simulation/de/rollercoaster/graphics/ViewObserver.java
@@ -1,5 +1,5 @@
 package de.rollercoaster.graphics;
 
 public interface ViewObserver {
-  public void update(float timeStep);
+  public void update(double timeStep);
 }
